Defining the ISchedule class as responsible for holding the information about a sync schedule (several sources, one destination for all) and a syncschedule manager for holding this information.
Defined WCF operations for managing ISchedule

Now on the server side, use a class to check for pending schedules and put them in the queue.

On the GUI side, create ISchedule objects to and from the viewmodel. After that, define a main screen, access to the history of syncs, current syncs and status, and the application should be completed